+++ crates/query/src/plan/orderby.rs
@@ -338,7 +338,7 @@ impl PlanNode for OrderByNode {
 
         obj.insert(
             "iterations".to_string(),
-            serde_json::json!(self.exec_info.iterations as u64),
+            serde_json::json!(self.exec_info.iterations),
         );
 
         let child_explain = self.source.explain_execute();
